extract settings queries out of database.ts
